requirements.txt
discord.py>=1.6
psutil>=5.7.3
pygame>=2.0.0
numpy>=1.20.0
pygame_gui>=0.5.7
psycopg2-binary>=2.8